youtubeData/requirements.txt
flask==1.0.3
flask_restplus==0.12.1
flask_cors==3.0.7
pyyaml==5.1
requests==2.22.0
youtube_transcript_api==0.1.4
flask_basicauth==0.2.0
pymongo==3.8.0
dnspython==1.16.0